# update xxhash and use Sum64String to avoid allocs